requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
psutil==5.9.8
python-dotenv==1.0.1
python-vlc==3.0.20123
requests==2.31.0
urllib3==2.2.1
